requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.26.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
jsonlines>=3.1.0
openai>=1.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
qwen-vl-utils>=0.0.2
